// frontend/src/api/generated/models/tender-create.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SENY Tender Automation
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InitialMaxPrice } from './initial-max-price';
// May contain unused imports in some cases
// @ts-ignore
import type { TenderSource } from './tender-source';

export interface TenderCreate {
    'number': string;
    'title': string;
    'customer': string;
    'description'?: string | null;
    'source'?: TenderSource;
    'source_url'?: string | null;
    'deadline_at': string;
    'published_at'?: string | null;
    'initial_max_price'?: InitialMaxPrice | null;
    'currency'?: string;
    'terms'?: { [key: string]: any; } | null;
    'responsible_id'?: number | null;
    'engineer_id'?: number | null;
}