// frontend/src/api/generated/models/tender-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SENY Tender Automation
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AuditLogResponse } from './audit-log-response';
// May contain unused imports in some cases
// @ts-ignore
import type { PositionResponse } from './position-response';
// May contain unused imports in some cases
// @ts-ignore
import type { StageResponse } from './stage-response';
// May contain unused imports in some cases
// @ts-ignore
import type { TenderFileResponse } from './tender-file-response';
// May contain unused imports in some cases
// @ts-ignore
import type { TenderSource } from './tender-source';

export interface TenderResponse {
    'number': string;
    'title': string;
    'customer': string;
    'description'?: string | null;
    'source'?: TenderSource;
    'source_url'?: string | null;
    'deadline_at': string;
    'published_at'?: string | null;
    'initial_max_price'?: string | null;
    'currency'?: string;
    'terms'?: { [key: string]: any; } | null;
    'responsible_id'?: number | null;
    'engineer_id'?: number | null;
    'id': number;
    'stage_id': number;
    'stage'?: StageResponse | null;
    'positions'?: Array<PositionResponse>;
    'files'?: Array<TenderFileResponse>;
    'audit_logs'?: Array<AuditLogResponse>;
    'is_archived': boolean;
    'created_at': string;
    'updated_at': string;
}